% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{data_model_normal}
\alias{data_model_normal}
\title{Normal data model}
\usage{
data_model_normal(prior_mean = 0, prior_sd = 1)
}
\arguments{
\item{prior_mean}{Prior mean for white noise standard deviation}

\item{prior_sd}{Prior standard deviation for white noise standard deviation}
}
\value{
lifeplus_data_model
}
\description{
Normal data model
}
\details{
The residuals are modeled as \eqn{\epsilon_{c,t} \sim N(0, \sigma^2)}.
The prior for the residual scale parameter is \eqn{\sigma \sim N(m, s^2)},
where \eqn{m} and \eqn{s} are user-specified via the \code{prior_mean} and \code{prior_sd} arguments, respectively.
}
