% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{data_model_outlier}
\alias{data_model_outlier}
\title{Outlier data model}
\usage{
data_model_outlier(outlier_threshold = 5, prior_mean = 0, prior_sd = 1)
}
\arguments{
\item{outlier_threshold}{threshold for ignoring observed life expectancy differences.}

\item{prior_mean}{Prior mean for white noise standard deviation}

\item{prior_sd}{Prior standard deviation for white noise standard deviation}
}
\value{
lifeplus_data_model
}
\description{
Outlier data model
}
