% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.lifeplus.R
\name{glance.lifeplus}
\alias{glance.lifeplus}
\title{Glance at a Lifeplus Model Fit}
\usage{
\method{glance}{lifeplus}(
  x,
  component = c("life", "shockfree"),
  conf.level = 0.95,
  diagnostics = TRUE,
  validations = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{lifeplus}, as returned by \code{\link{lifeplus}()}.}

\item{component}{A character string specifying which variable to use for validations.
One of \code{"life"} (default), \code{"shockfree"}. The latter
requires the presence of a shock model.}

\item{conf.level}{A scalar in \code{(0, 1)} specifying the credible interval width.
Defaults to \code{0.95} (i.e., a 95\% interval).}

\item{diagnostics}{Logical; whether to include MCMC diagnostic summary measures (default: \code{TRUE})}

\item{validations}{Logical; whether to include held-out data validation summary measures (default: \code{TRUE})}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A \code{\link[tibble]{tibble}} with columns:
\describe{
\item{\code{transition.model}}{Transition model used.}
\item{\code{shock.model}}{Shock model used.}
\item{\code{data.model}}{Data model used.}
\item{\code{num_divergent}}{Number of divergent transitions (total across all chains).}
\item{\code{num_max_treedepth}}{Number of transitions that hit max treedepth (total across all chains).}
\item{\code{min_ebfmi}}{Minimum E-BFMI across all chains.}
\item{\code{mean.error}}{Mean error.}
\item{\code{median.error}}{Median error.}
\item{\code{mean.absolute.error}}{Mean absolute error.}
\item{\code{median.absolute.error}}{Median absolute error.}
\item{\code{ci.width}}{Width of credible interval of level specified by \code{conf.level}.}
\item{\code{coverage}}{Empirical coverage of the credible interval of level specified by \code{conf.level}.}
\item{\code{conf.level}}{Credible interval level specified by \code{conf.level}}
}
}
\description{
Extracts principle diagnostic and validation results from a Lifeplus model fit
following the conventions of the \code{broom} package.
Each row represents one fit. If the model fit includes held-out observations,
model validation summaries are extracted based on the most recent held-out observations in each area.
}
\seealso{
\code{\link{lifeplus}()}, \code{\link{summary.lifeplus}()}
}
