% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeplus.R
\name{lifeplus}
\alias{lifeplus}
\title{Fit Lifeplus model}
\usage{
lifeplus(
  data,
  y,
  time,
  area,
  start_time = NULL,
  end_time = NULL,
  transition_model = transition_model_logistic(),
  data_model = data_model_normal(),
  shock_model = shock_model_none(),
  held_out = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{y}{column name of outcome variable..}

\item{time}{column name of time variable.}

\item{area}{column name of the area of each observation}

\item{start_time}{start time of estimates.}

\item{end_time}{end time of estimates.}

\item{transition_model}{transition model specification (default: \code{transition_model_logistic()})}

\item{data_model}{data model specification (default: \code{data_model_normal()})}

\item{shock_model}{shock term specification (default: \code{shock_model_none()})}

\item{held_out}{binary vector indicating which observations are held out. Set to FALSE to hold out no observations.}

\item{...}{additional arguments passed to CmdStanModel::sample.}
}
\description{
Fit Lifeplus model
}
