% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavity.R
\name{lifeplus_cavity_prior}
\alias{lifeplus_cavity_prior}
\title{Cavity prior specification for expectation propagation.}
\usage{
lifeplus_cavity_prior(mu, Sigma)
}
\arguments{
\item{mu}{numeric vector specifying cavity prior mean}

\item{Sigma}{numeric matrix specifying cavity prior covariance matrix}
}
\value{
object of class \code{lifeplus_cavity_prior}
}
\description{
Cavity prior specification for expectation propagation.
}
