% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lifeplus_default_posterior_quantiles}
\alias{lifeplus_default_posterior_quantiles}
\title{Default summary posterior quantile probabilities}
\usage{
lifeplus_default_posterior_quantiles()
}
\value{
numeric vector of quantile probabilities
}
\description{
Default summary posterior quantile probabilities
}
\details{
The default quantiles are 0.1\%, 1\%, 2.5\%, 25\%, 50\%, 75\%, 90\%, 97.5\%, 99\%, 99.9\%.
}
