% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeplus_influence.R
\name{lifeplus_influence}
\alias{lifeplus_influence}
\title{Compute posterior covariance influence measure}
\usage{
lifeplus_influence(x, parameters)
}
\arguments{
\item{x}{An object of class \code{lifeplus}, as returned by \code{\link{lifeplus}()}.}

\item{parameters}{Vector of model parameter names}
}
\value{
vector of influence measures
}
\description{
Computes the posterior covariance between the log likelihood and the given parameter.
}
