% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_projections}
\alias{plot_projections}
\title{Plot posterior projections}
\usage{
plot_projections(
  fit,
  areas = fit$areas,
  start_time = min(fit$times),
  end_time = max(fit$times),
  data = TRUE
)
}
\arguments{
\item{fit}{lifeplus fit object}

\item{areas}{vector of areas to plot (default: all areas in fit)}

\item{start_time}{first time point to plot (defaults to first data point)}

\item{end_time}{final time point to plot (defaults to final projection time point)}

\item{data}{logical indicating whether to plot raw data (default: TRUE)}
}
\value{
ggplot2 plot
}
\description{
Plot posterior projections
}
