% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_shocks}
\alias{plot_shocks}
\title{Plot posterior shocks}
\usage{
plot_shocks(
  fit,
  areas = fit$areas,
  start_time = min(fit$times),
  end_time = max(fit$times)
)
}
\arguments{
\item{fit}{lifeplus fit object}

\item{areas}{vector of areas to plot (default: all areas in fit)}

\item{start_time}{first time point to plot (defaults to first data point)}

\item{endart_time}{final time point to plot (defaults to final projection time point)}
}
\description{
Plot posterior shocks
}
