% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_transitions}
\alias{plot_transitions}
\title{Plot posterior transition functions}
\usage{
plot_transitions(fit, areas = fit$areas, data = TRUE)
}
\arguments{
\item{fit}{lifeplus fit object}

\item{areas}{vector of areas to plot (default: all areas in fit)}

\item{data}{logical indicating whether to plot raw data (default: TRUE)}
}
\value{
ggplot2 object
}
\description{
Plot posterior transition functions
}
