% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_validation}
\alias{process_validation}
\title{Estimates validation metrics for held-out data}
\usage{
process_validation(fit)
}
\arguments{
\item{fit}{a (partially constructed) fit object}
}
\description{
Estimates validation metrics for held-out data
}
