% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R, R/shock_models.R
\name{shock_model_none}
\alias{shock_model_none}
\title{Fix all shock terms to zero.}
\usage{
shock_model_none()

shock_model_none()
}
\description{
Fix all shock terms to zero.

Fix all shock terms to zero.
}
