% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lifeplus.R
\name{simulate_lifeplus}
\alias{simulate_lifeplus}
\title{Generate simulated data for testing lifeplus models.}
\usage{
simulate_lifeplus(num_areas, num_times, seed = NULL)
}
\arguments{
\item{num_areas}{number of areas to simulate}

\item{num_times}{number of time points to simulate}

\item{seed}{optional random number seed}
}
\value{
data frame
}
\description{
Generate simulated data for testing lifeplus models.
}
