% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lifeplus.R
\name{summary.lifeplus}
\alias{summary.lifeplus}
\title{Summarise a Lifeplus Model Fit}
\usage{
\method{summary}{lifeplus}(
  object,
  component = c("life", "shocks", "shockfree", "transition_functions",
    "transition_function_mean", "transition_params", "all"),
  areas = NULL,
  times = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{lifeplus}, as returned by \code{\link{lifeplus}()}.}

\item{component}{A character string specifying which posterior component to summarise: one of
"life", "shocks", "shockfree", transition_functions", "transition_function_mean", "transition_params", or "all".}

\item{areas}{An optional character vector of areas names to include in summary.}

\item{times}{An optional integer vector of time points to include in summary.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class \code{summary.lifeplus}.
}
\description{
Produces a structured summary of posterior estimates from a fitted \code{lifeplus} model.
}
\seealso{
\code{link{lifeplus}()}, \code{\link{tidy.lifeplus}()}, \code{\link{print.lifeplus}()}
}
