% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{transition_model_spline}
\alias{transition_model_spline}
\title{Spline transition model}
\usage{
transition_model_spline(degree = 2, num_knots = 7)
}
\arguments{
\item{degree}{Spline degree}

\item{num_knots}{Number of spline knots}
}
\value{
lifeplus_transition_model
}
\description{
Spline transition model
}
