% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{data_model_outlier}
\alias{data_model_outlier}
\title{Outlier data model}
\usage{
data_model_outlier(outlier_threshold = 5, prior_mean = 0, prior_sd = 1)
}
\arguments{
\item{outlier_threshold}{threshold for ignoring observed life expectancy differences.}

\item{prior_mean}{Prior mean for white noise standard deviation}

\item{prior_sd}{Prior standard deviation for white noise standard deviation}
}
\value{
lifeplus_data_model
}
\description{
Outlier data model
}
\details{
Any observed absolute differences in life expectancy greater than the user-specified outlier threshold (\code{outlier_threshold}) are ignored.
The remaining residuals are modeled as in the normal data model:  $\epsilon_{c,t} \sim N(0, \sigma^2)$.
The prior for the residual scale parameter is $\sigma \sim N(m, s^2)$,
where $m$ and $s$ are user-specified via the \code{prior_mean} and \code{prior_sd} arguments, respectively.
}
