% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{shock_model_regularized_horseshoe}
\alias{shock_model_regularized_horseshoe}
\title{Regularized horseshoe prior model for shocks}
\usage{
shock_model_regularized_horseshoe(
  scale_global = 0.1,
  slab_scale = 10,
  slab_df = 6,
  constrain_negative = FALSE
)
}
\arguments{
\item{scale_global}{Global scale parameter}

\item{slab_scale}{Slab scale parameter}

\item{slab_df}{Slab degrees of freedom parameter}

\item{constrain_negative}{Whether to constrain shocks to be negative (default: FALSE)}
}
\description{
Regularized horseshoe prior model for shocks
}
