% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock_models.R
\name{shock_model_regularized_horseshoe}
\alias{shock_model_regularized_horseshoe}
\title{Regularized horseshoe prior model for shock terms.}
\usage{
shock_model_regularized_horseshoe(
  scale_global = 0.1,
  slab_scale = 10,
  slab_df = 6,
  constrain_negative = FALSE
)
}
\arguments{
\item{scale_global}{Global scale parameter}

\item{slab_scale}{Slab scale parameter}

\item{slab_df}{Slab degrees of freedom parameter}

\item{constrain_negative}{Whether to constrain shocks to be negative (default: FALSE)}
}
\description{
Regularized horseshoe prior model for shock terms.
}
\details{
The regularized horseshoe prior aggressively shrinks the shock term to zero.
}
\references{
Juho Piironen, Aki Vehtari "Sparsity information and regularization in the horseshoe and other shrinkage priors,"
Electronic Journal of Statistics, Electron. J. Statist. 11(2), 5018-5051, (2017)
}
