% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{transition_model_double_logistic}
\alias{transition_model_double_logistic}
\title{Double logistic transition model}
\usage{
transition_model_double_logistic(hierarchical = TRUE)
}
\arguments{
\item{hierarchical}{whether to estimate double logistic parameters hierarchically (default: TRUE)}
}
\value{
Object of class lifeplus_transition_model
}
\description{
Double logistic transition model
}
