% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_models.R
\name{transition_model_gaussian_process}
\alias{transition_model_gaussian_process}
\title{Approximate Gaussian Process transition model}
\usage{
transition_model_gaussian_process(
  hierarchical = TRUE,
  basis_functions = 20,
  boundary_factor = 1.5
)
}
\arguments{
\item{hierarchical}{Whether to estimate hierarchical Gaussian Process}

\item{basis_functions}{Number of basis functions to use in approximation}

\item{boundary_factor}{Approximation boundary factor tuning parameter}
}
\value{
lifeplus_transition_model
}
\description{
Approximate Gaussian Process transition model
}
\details{
The transition function is modeled by a Gaussian Process prior.
For computational reasons, the Gaussian Process is approximated with a basis function approach
(see Riutort-Mayol et al. 2020).

References: Riutort-Mayol, G., Bürkner, PC., Andersen, M.R. et al. (2020) Practical Hilbert space approximate Bayesian Gaussian processes for probabilistic programming. Stat Comput 33, 17
}
