% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R, R/transition_models.R
\name{transition_model_spline}
\alias{transition_model_spline}
\title{Spline transition model}
\usage{
transition_model_spline(hierarchical = TRUE, degree = 2, knots = 7)

transition_model_spline(hierarchical = TRUE, degree = 2, knots = 7)
}
\arguments{
\item{hierarchical}{Whether to apply hierarchical models to spline coefficients}

\item{degree}{Spline degree}

\item{knots}{Number of spline knots}

\item{num_knots}{Number of spline knots}
}
\value{
lifeplus_transition_model

lifeplus_transition_model
}
\description{
Spline transition model

Spline transition model
}
